project-06-data-pipeline-sentinel/pyproject.toml
[project]
name = "data-pipeline-sentinel"
version = "0.1.0"
description = "LangGraph data pipeline sentinel with episodic memory and Great Expectations"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "great-expectations>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "kafka-python-ng>=2.2.0",
    "boto3>=1.34.0",
    "PyGithub>=2.3.0",
    "asyncpg>=0.29.0",
    "chromadb>=0.6.0",
    "openai>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
